Testing and hw updated
